Add files via upload
Add findmineven function

Implemented a Python function, findmineven, to find the smallest even number in a given sequence. The function takes a list of integers as input and returns the smallest even number. If no even number is found, it returns -1.
The function has been tested with a sample sequence in the provided driver code.